refactor(provision): reorganize code, add logging
